enable selection of different LLM clients
